+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 Nothing yet.
 
+## 1.2.0 - Fri 29th Jan 2016
+
+### New
+
+* Added support for extracting the first entry from a list
+  - Added `Filters\ExtractFirstItem`
+
 ## 1.1.0 - Thu 28th Jan 2016
 
 ### New